backend/requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-filter==25.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.3.17
langchain-community==0.3.24
pyyaml==6.0.2
markdown==3.8
gunicorn==21.2.0

# File processing dependencies
python-docx==1.0.1
PyPDF2==3.0.1
textract==1.6.5